regras para distribuicao das pontuacoes

+++ lineup/src/main/java/com/sevenine/lineup/application/usecase/input/PointInput.java
@@ -3,6 +3,5 @@ package com.sevenine.lineup.application.usecase.input;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record PointInput(double assist, double block, double save, double shot, double goal, double pass,
-                         double punch, double foul) {
+public record PointInput(int assist, int block, int save, int shot, int goal, int pass, int punch, int foul) {
 }
